% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerDueDays.R
\name{customerDueDays_upload}
\alias{customerDueDays_upload}
\title{客户账期上传}
\usage{
customerDueDays_upload(token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
客户账期上传
}
\examples{
customerDueDays_upload()
}
