% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerDueDays.R, R/customerDueDays2.R
\name{customerDueDays_view}
\alias{customerDueDays_view}
\title{客户账期查询}
\usage{
customerDueDays_view(token)

customerDueDays_view(token)
}
\arguments{
\item{token}{}
}
\value{



}
\description{
客户账期查询

客户账期查询
}
\examples{
customerDueDays_view()
customerDueDays_view()
}
