% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackBill.R
\name{blackBill_delete}
\alias{blackBill_delete}
\title{黑名单删除}
\usage{
blackBill_delete(token, FProjectNumber)
}
\arguments{
\item{FProjectNumber}{}
}
\value{

}
\description{
黑名单删除
}
\examples{
blackBill_delete()
}
