% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerDueDays.R
\name{customerDueDays_delete}
\alias{customerDueDays_delete}
\title{客户账期删除}
\usage{
customerDueDays_delete(token, FCustomerNumber)
}
\arguments{
\item{FCustomerNumber}{}
}
\value{

}
\description{
客户账期删除
}
\examples{
customerDueDays_delete()
}
